updated to be compatible with Neos.Flow v4.0
