+++ Classes/Service/BindProvider/LdapBind.php
@@ -1,8 +1,8 @@
 <?php
-namespace TYPO3\LDAP\Service\BindProvider;
+namespace Neos\Ldap\Service\BindProvider;
 
 /*                                                                        *
- * This script belongs to the Flow package "TYPO3.LDAP".                  *
+ * This script belongs to the Flow package "Neos.Ldap".                  *
  *                                                                        *
  * It is free software; you can redistribute it and/or modify it under    *
  * the terms of the GNU Lesser General Public License as published by the *
@@ -18,18 +18,18 @@ namespace TYPO3\LDAP\Service\BindProvider;
  * License along with the script.                                         *
  * If not, see http://www.gnu.org/licenses/lgpl.html                      *
  *                                                                        *
- * The TYPO3 project - inspiring people to share!                         *
+ * The Neos project - inspiring people to share!                         *
  *                                                                        */
 
-use TYPO3\Flow\Annotations as Flow;
-use TYPO3\Flow\Error\Exception;
+use Neos\Flow\Annotations as Flow;
+use Neos\Flow\Error\Exception;
 
 /**
  * Bind to an OpenLDAP Server
  *
  * @Flow\Scope("prototype")
  */
-class LDAPBind extends AbstractBindProvider
+class LdapBind extends AbstractBindProvider
 {
 
     /**
