+++ Configuration/Settings.yaml
@@ -1,5 +1,5 @@
-TYPO3:
-  LDAP:
+Neos:
+  Ldap:
     # Allow login based on cached credentials if the directory service is offline
     allowStandinAuthentication: FALSE
     roles:
