+++ composer.json
@@ -1,15 +1,18 @@
 {
-	"name": "typo3/ldap",
-	"type": "typo3-flow-package",
+	"name": "neos/ldap",
+	"type": "neos-package",
 	"description": "LDAP Authentication for Flow",
 	"license": ["LGPL-3.0+"],
 	"require":{
-		"typo3/flow": "*",
+		"neos/flow": "^4.0",
 		"ext-ldap": "*"
 	},
+	"replace": {
+		"typo3/ldap": "self.version"
+	},
 	"autoload":{
-		"psr-0":{
-			"TYPO3\\LDAP": "Classes"
+		"psr-4":{
+			"Neos\\Ldap\\": "Classes"
 		}
 	}
 }
